server

+++ src/core/game.js
@@ -1,5 +1,5 @@
 import Serializable from './serializable'
-import '../components/import'
+import { addChange, changeType, setChangeOrigin } from './serializableManager';
 import Prototype from '../core/prototype'
 import assert from '../util/assert'
 import PropertyOwner, { Prop } from '../core/propertyOwner';
@@ -8,18 +8,23 @@ let propertyTypes = [
 	Prop('name', 'No name', Prop.string)
 ];
 
-export let game = null; // only one game at the time
+let game = null; // only one game at the time
+export { game };
+
+let isClient = typeof window !== 'undefined';
 
 export default class Game extends PropertyOwner {
 	constructor(predefinedId) {
-		if (game) {
-			try {
-				game.delete();
-			} catch(e) {
-				console.warn('Deleting old game failed', e);
+		if (isClient) {
+			if (game) {
+				try {
+					game.delete();
+				} catch (e) {
+					console.warn('Deleting old game failed', e);
+				}
 			}
+			game = this;
 		}
-		game = this;
 		
 		if (predefinedId)
 			console.log('game import');
@@ -27,6 +32,10 @@ export default class Game extends PropertyOwner {
 			console.log('game created');
 		
 		super(...arguments);
+	}
+	initWithChildren() {
+		super.initWithChildren(...arguments);
+		addChange(changeType.addSerializableToTree, this);
 	}
 	delete() {
 		if (!super.delete()) return false;
